finalizando as telas

+++ EmprestaGames.Api/Respositories/JogoRepository.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using EmprestaGames.Api.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace EmprestaGames.Api.Respositories
 {
@@ -43,7 +44,21 @@ namespace EmprestaGames.Api.Respositories
                 using (EGContext db = new EGContext())
                 {
                     var jogo = model;
-                    db.Jogos.Add(model);                    
+                    if (model.Id > 0)
+                    {
+                        var exists = db.Jogos.Find(model.Id);
+                        if (exists == null) return null;
+                        exists.Descricao = model.Descricao;
+                        exists.Nome = model.Nome;
+                        db.Entry(exists).State = EntityState.Modified;
+
+                    }
+                    else
+                    {
+                        db.Jogos.Add(model);
+                    }
+                    
+                                       
                     db.SaveChanges();
 
                     return jogo;
@@ -55,13 +70,15 @@ namespace EmprestaGames.Api.Respositories
             }
         }
 
-        public bool Remove(Jogo model)
+        public bool Remove(int Id)
         {
             try
             {
                 using (EGContext db = new EGContext())
                 {
-                    
+                    if (Id == 0) return false;
+
+                    var model = db.Jogos.Find(Id);
                     db.Jogos.Remove(model);                    
                     db.SaveChanges();
