+++ EmprestaGames.Api/Respositories/PessoaRepository.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using EmprestaGames.Api.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace EmprestaGames.Api.Respositories
 {
@@ -43,11 +44,20 @@ namespace EmprestaGames.Api.Respositories
                 using (EGContext db = new EGContext())
                 {
                     var pessoa = model;
-                    db.Pessoas.Add(model);                    
+                    if (model.Id > 0)
+                    {
+                        var exists = db.Pessoas.Find(model.Id);
+                        if (exists == null) return null;                        
+                        exists.Nome = model.Nome;
+                        db.Entry(exists).State = EntityState.Modified;
+
+                    }
+                    else
+                    {
+                        db.Pessoas.Add(pessoa);
+                    }
+
                     db.SaveChanges();
-
-
-
                     return pessoa;
                 }
             }
@@ -57,13 +67,15 @@ namespace EmprestaGames.Api.Respositories
             }
         }
 
-        public bool Remove(Pessoa model)
+        public bool Remove(int Id)
         {
             try
             {
                 using (EGContext db = new EGContext())
                 {
-                    
+                    if (Id == 0) return false;
+
+                    var model = db.Pessoas.Find(Id);
                     db.Pessoas.Remove(model);                    
                     db.SaveChanges();
